+++ src/main/java/com/braintreegateway/customersession/InsightPaymentOption.java
@@ -1,0 +1,6 @@
+package com.braintreegateway.customersession;
+
+public enum InsightPaymentOption {
+    PAYPAL,
+    VENMO
+}