+++ src/main/java/com/braintreegateway/customersession/Insights.java
@@ -1,0 +1,5 @@
+package com.braintreegateway.customersession;
+
+public enum Insights {
+    PAYMENT_INSIGHTS
+}