+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Add `blikAliases` to LocalPaymentCompleted webhook and LocalPaymentDetails
 * Deprecate `SamsungPayCard`
 * Updated expiring pinned vendor SSL certificates
+* Add support for creating and updating PayPal customer session
+* Add support for getting PayPal customer insights
 
 ## 3.36.0
 * Add `payerName`, `bic` and `iban_last_chars` for LocalPaymentCompleted webhook